refactor: :truck: Move component files and their stylesheets to components/{componentName}.
Index.ts imports and exports all components for easier access.
