+++ components/Main/Main.tsx
@@ -1,4 +1,4 @@
-import styles from "../styles/Main.module.css";
+import styles from "./Main.module.css";
 
 export default function Main(): JSX.Element {
 	return (
