+++ components/index.ts
@@ -1,0 +1,6 @@
+import About from "./About/About";
+import Header from "./Header/Header";
+import Main from "./Main/Main";
+import Portfolio from "./Portfolio/Portfolio";
+
+export { About, Header, Main, Portfolio };
