+++ pages/_app.tsx
@@ -1,5 +1,5 @@
 import { AppProps } from "next/app";
-import Header from "../components/Header";
+import { Header } from "../components/index";
 import "../styles/globals.css";
 
 export default function MyApp({ Component, pageProps }: AppProps): JSX.Element {
