+++ pages/index.tsx
@@ -1,7 +1,5 @@
 import { DetailedHTMLProps, HTMLAttributes } from "react";
-import Main from "../components/Main";
-import About from "../components/About";
-import Portfolio from "../components/Portfolio";
+import { Main, About, Portfolio } from "../components/index";
 
 export default function Home(): DetailedHTMLProps<
 	HTMLAttributes<HTMLDivElement>,
